* First cut at Solaris fixes. There may still be some *printf("%s", NULL)    type of issues lurking in the shadows.

+++ citadel/rcit.c
@@ -34,7 +34,7 @@ int main(int argc, char **argv)
 	char spool_only = 0;
 
 	get_config();
-	sprintf(flnm,"./network/spoolin/rcit.%d", getpid());
+	sprintf(flnm,"./network/spoolin/rcit.%ld", (long)getpid());
 	sprintf(tname, tmpnam(NULL));
 
 	for (a=1; a<argc; ++a) {
